requirements.txt
flask
flask-cors
gunicorn
waitress
requests
python-dotenv
